Split of single and multiselect components
